pyproject.toml
[tool.poetry]
name = "text-corpus-creation"
version = "0.1.0"
description = "Text corpus creation tools for remote sensing papers"
packages = [
    { include = "causal_llm" },
    { include = "diffusion_model" },
    { include = "downstream_evaluation" },
    { include = "downstream_data_generation" },
    { include = "inference" },
    { include = "metrics" },
    { include = "utils" }
]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.0"
pypdf = "^3.0.0"
gradio = "^4.0.0"
torch = ">=2.0.0,<3.0.0"
diffusers = "^0.25.0"
transformers = "^4.0.0"
pillow = "^10.0.0"
datasets = "^2.0.0"
bitsandbytes = ">=0.41.1,<0.42.0"
peft = "^0.7.0"
torchmetrics = "0.11.4"
torchvision = ">=0.15.0,<1.0.0"
numpy = "^1.24.0"
accelerate = "^0.25.0"
evaluate = "^0.4.0"
huggingface-hub = "^0.24.0"
pyarrow = "^14.0.1"
pytorch-lightning = "^2.1.2"
albumentations = "^1.3.1"
torchgeo = "0.5.0"
timm = "0.9.7"
segmentation-models-pytorch = "0.3.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate-corpus = "text_corpus_creation.text_data_generation_pipeline:main"
launch-app = "inference.app:main"
train-model = "text_corpus_creation.train:main"
calculate-fid = "metrics.metric:main"
format-captions = "diffusion_model.utils:main"
generate-data = "downstream_data_generation.generate:main"
evaluate-model = "downstream_evaluation.downstream:main" 